+++ app_state/example/lib/6_forward_recovery/pages/input/path.dart
@@ -5,7 +5,7 @@ import 'page.dart';
 import '../home/path.dart';
 
 class InputPath extends PagePath {
-  static const _location = '/input';
+  static final _url = Uri.parse('/input');
 
   final String text;
 
@@ -18,10 +18,10 @@ class InputPath extends PagePath {
         );
 
   @override
-  get location => _location;
+  Uri get uri => _url;
 
   static InputPath? tryParse(RouteInformation ri) {
-    return ri.location == _location ? InputPath(text: '') : null;
+    return ri.uri == _url ? InputPath(text: '') : null;
   }
 
   @override