+++ app_state/example/lib/7_any_tab/pages/tab2/path.dart
@@ -4,15 +4,15 @@ import 'package:flutter/widgets.dart';
 import '../../router/tab_enum.dart';
 
 class Tab2Path extends PagePath {
-  static const _location = '/two';
+  static final _url = Uri.parse('/two');
 
   const Tab2Path() : super(key: 'Tab2');
 
   @override
-  String get location => _location;
+  Uri get uri => _url;
 
   static Tab2Path? tryParse(RouteInformation ri) {
-    return ri.location == _location ? const Tab2Path() : null;
+    return ri.uri == _url ? const Tab2Path() : null;
   }
 
   @override