+++ app_state/example/lib/9_state/pages/home/path.dart
@@ -1,10 +1,10 @@
 import 'package:app_state/app_state.dart';
 
 class HomePath extends PagePath {
-  static const _location = '/';
+  static final _url = Uri.parse('/');
 
   const HomePath() : super(key: 'Home');
 
   @override
-  String get location => _location;
+  Uri get uri => _url;
 }